+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/V20250203/Controllers/CertificatesController.cs
@@ -2,7 +2,7 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using TeachingRecordSystem.Api.Infrastructure.Security;
 
-namespace TeachingRecordSystem.Api.V3.VNext.Controllers;
+namespace TeachingRecordSystem.Api.V3.V20250203.Controllers;
 
 [Route("certificates")]
 [Authorize(AuthorizationPolicies.IdentityUserWithTrn)]
