+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/V20250203/Requests/FindPersonRequest.cs
@@ -1,6 +1,6 @@
 using Microsoft.AspNetCore.Mvc;
 
-namespace TeachingRecordSystem.Api.V3.VNext.Requests;
+namespace TeachingRecordSystem.Api.V3.V20250203.Requests;
 
 public record FindPersonRequest
 {
