+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/V20250203/Requests/FindPersonsRequest.cs
@@ -1,4 +1,4 @@
-namespace TeachingRecordSystem.Api.V3.VNext.Requests;
+namespace TeachingRecordSystem.Api.V3.V20250203.Requests;
 
 public record FindPersonsRequest
 {
