+++ TeachingRecordSystem/src/TeachingRecordSystem.Api/V3/V20250203/Responses/FindPersonsResponse.cs
@@ -1,22 +1,23 @@
+using AutoMapper.Configuration.Annotations;
 using TeachingRecordSystem.Api.V3.Implementation.Operations;
-using TeachingRecordSystem.Api.V3.VNext.Requests;
-using TeachingRecordSystem.Core.ApiSchema.V3.V20240101.Dtos;
 using TeachingRecordSystem.Core.ApiSchema.V3.V20240814.Dtos;
 using TeachingRecordSystem.Core.ApiSchema.V3.V20240920.Dtos;
-using InductionStatus = TeachingRecordSystem.Core.ApiSchema.V3.VNext.Dtos.InductionStatus;
-using QtlsStatus = TeachingRecordSystem.Core.ApiSchema.V3.VNext.Dtos.QtlsStatus;
+using InductionStatus = TeachingRecordSystem.Core.ApiSchema.V3.V20250203.Dtos.InductionStatus;
+using NameInfo = TeachingRecordSystem.Core.ApiSchema.V3.V20240101.Dtos.NameInfo;
+using QtlsStatus = TeachingRecordSystem.Core.ApiSchema.V3.V20250203.Dtos.QtlsStatus;
 
-namespace TeachingRecordSystem.Api.V3.VNext.Responses;
+namespace TeachingRecordSystem.Api.V3.V20250203.Responses;
 
-public record FindPersonResponse
+[AutoMap(typeof(FindPersonsResult))]
+public record FindPersonsResponse
 {
     public required int Total { get; init; }
-    public required FindPersonRequest Query { get; init; }
-    public required IReadOnlyCollection<FindPersonResponseResult> Results { get; init; }
+    [SourceMember(nameof(FindPersonsResult.Items))]
+    public required IReadOnlyCollection<FindPersonsResponseResult> Results { get; init; }
 }
 
 [AutoMap(typeof(FindPersonsResultItem))]
-public record FindPersonResponseResult
+public record FindPersonsResponseResult
 {
     public required string Trn { get; init; }
     public required DateOnly DateOfBirth { get; init; }
