+++ TeachingRecordSystem/src/TeachingRecordSystem.Core/ApiSchema/V3/V20250203/Dtos/InductionInfo.cs
@@ -1,4 +1,4 @@
-namespace TeachingRecordSystem.Core.ApiSchema.V3.VNext.Dtos;
+namespace TeachingRecordSystem.Core.ApiSchema.V3.V20250203.Dtos;
 
 public record InductionInfo
 {
