+++ TeachingRecordSystem/tests/TeachingRecordSystem.Api.Tests/V3/V20250203/GetPersonTests.cs
@@ -1,6 +1,6 @@
 using TeachingRecordSystem.Api.V3.Implementation.Dtos;
 
-namespace TeachingRecordSystem.Api.Tests.V3.VNext;
+namespace TeachingRecordSystem.Api.Tests.V3.V20250203;
 
 public class GetPersonTests(HostFixture hostFixture) : TestBase(hostFixture)
 {
@@ -36,8 +36,7 @@ public class GetPersonTests(HostFixture hostFixture) : TestBase(hostFixture)
             {
                 status = status.ToString(),
                 startDate = startDate.ToString("yyyy-MM-dd"),
-                completedDate = completedDate.ToString("yyyy-MM-dd"),
-                certificateUrl = "/v3/certificates/induction"
+                completedDate = completedDate.ToString("yyyy-MM-dd")
             },
             responseInduction);
     }
@@ -63,8 +62,7 @@ public class GetPersonTests(HostFixture hostFixture) : TestBase(hostFixture)
             {
                 status = InductionStatus.None.ToString(),
                 startDate = (DateOnly?)null,
-                completedDate = (DateOnly?)null,
-                certificateUrl = (string?)null
+                completedDate = (DateOnly?)null
             },
             responseInduction);
     }
